Drop a python 2 exception management
Previously (when python 2 was supported) we introduced some specific
exception management to detect connections issues when a rabbitmq
cluster node disappearing.

The original issue was that a ConnectionRefusedError is thrown and not
managed by oslo.messaging to detect heartbeat issue and call ensure_connection
for switching the connection destination (cluster node).

`ConnectionRefusedError` is only a python 3 exception [2].

Now that we only support python 3 we don't need to continue to
wrap/emulate this kind of exceptions (ConnectionRefusedError) [2] and
so we can drop the python 2 compatibility code to only support python 3 code.

[1] https://github.com/openstack/oslo.messaging/commit/9d8b1430e5c081b081c0e3c0b5f12f744dc7809d
[2] https://docs.python.org/3/library/exceptions.html#ConnectionRefusedError

Change-Id: I4c459d8c947dac213a1866c0d37e8f3d547aa82e