Python 3.9: isAlive is removed
The Threading method isAlive was deprecated a long time ago, in the favor
of is_alive(). Though in Python 3.9, isAlive is removed. We must switch
to it. Note that is_alive() is available at least in Python 3.5 (I tried)
and probably even earlier, so switching to is_alive() is not a problem
for the Python interpreter versions currently supported by OpenStack.

Change-Id: I9d671abcd2cea9c0c726edaddcd65e1093d96731