Fix type of direct_mandatory_flag opt
An IntOpt with a default of True is invalid. I'm a little surprised
this doesn't fail a defaults check somewhere, but it needs to be
fixed regardless.

Looking at where it is used, it appears the boolean type is correct.
This just changes the opt type to BoolOpt to match.

Change-Id: I01a38754a31c891f2b3b9c7f8135690693df5d13
Closes-Bug: 1909036