Disable greenthreads for RabbitDriver "listen" connections
When enabling heartbeat_in_pthread, we were restoring the "threading"
python library from eventlet to original one in RabbitDriver but we
forgot to do the same in AMQPDriverBase (RabbitDriver is subclass of
AMQPDriverBase).

We also need to use the original "queue" so that queues are not going to
use greenthreads as well.

Related-bug: #1961402
Related-bug: #1934937
Closes-bug: #2009138

Signed-off-by: Arnaud Morin <arnaud.morin@ovhcloud.com>
Change-Id: I34ea0d1381e934297df2f793e0d2594ef8254f00
(cherry picked from commit 3645839d162a989da46ba1c0fcbe7f32b48b9fb1)