Set a default value for metrics process_name
When instanciating a metrics client from oslo.messaging, we usually
need to set a metrics_process_name to uniquely identify the client.

The default value is now enforced from the process basename to ease
usage of such option.

This is very useful for e.g. when using multiple neutron clients on
computes (like l3, dhcp, metadata) that reads the same config files and
cannot have a different setting.

Change-Id: I16b26cfcc91a24ec6a14c84f11bf54e80a77ec0b
Signed-off-by: Arnaud Morin <arnaud.morin@ovhcloud.com>