Make oslo.messaging, magnum, and zaqar reproducible.
Whilst working on the Reproducible Builds effort [0], we noticed that
python-oslo.messaging could not be built reproducibly.

This is because the documentation captures the hostname of the build
system.

[0] https://reproducible-builds.org/

This patch uses sample_default from oslo.config to fix this.

Please accept this patch to fix oslo.messaging, magnum, and zaqar
(at least, probably others).

Change-Id: Ie8717e182f709cbddd645a356789e262b47646d3