[doc] Better document rabbit_transient_queues_ttl option
With implementation of rabbit_stream_fanout option,
rabbit_transient_queues_ttl is now used to control the size of stream
as well.
While it made sense to re-use the option, it is behaving quite
differently from what is currently described as now it controls not the
queue TTL itself, but messages TTL inside the queue, which is good to
be explicit about.

Change-Id: I78c5066d5d62192f2206893c04823b83791ebdc3