Fix queue_manager in a containerized environment
This patch fixes the operation of queue_manager in a
containerized environment by adding an additional check
on the start_time in ticks since boot.
This way, we can detect a restart even when the PID remains
unchanged as it is ussual in containers, but the start_time is
different.

[1] https://www.man7.org/linux/man-pages//man5/proc_pid_stat.5.html

From man page above:

(22) starttime  %llu
    The time the process started after system boot.
    Before Linux 2.6, this value was expressed in
    jiffies.  Since Linux 2.6, the value is expressed
    in clock ticks (divide by sysconf(_SC_CLK_TCK)).

Closes-Bug: #2078935
Change-Id: I9e22433ec039ad6783593d9cb7fbe22c9090534e
(cherry picked from commit 6790f702fa5a56c9d25a4becba670bf1810a4162)