Remove pyngus from dependencies
The `pyngus` library was introduced as a dependency for the AMQP 1.0 driver.
Since the AMQP 1.0 driver has been removed from the project, this library is
no longer needed. This patch cleans up the dependencies by removing `pyngus`
from the requirements.

A dependency check (`pipdeptree`) confirms that `pyngus` is not used directly
or indirectly in the project.

Change-Id: Idf94ceaa96005a74e60fafc0124df659a8854ea5