Fixed: use self.start_time in QManager shm
In a previous patch, start_time was introduced to take care of
containerized services.
Unfortunately, a bug has been introduced during this change.

The start_time written in /dev/shm was using the previous start_time
written in the file instead of the new one computed at boot of the
service.

Change-Id: I8a31c0c38b42f333bba92159332e7b3b79d70854
Signed-off-by: Arnaud Morin <arnaud.morin@ovhcloud.com>