Do not delete queues on error
When a queue is in error, we used to delete the faulty queue.
The problem with that code is that under heavy load, for queues that are
shared accross agents, this could fail in a loop and never recover.

Agent A detect the queue is broken:
it delete the queue and then recreate

In the meantime, Agent B detect the same thing:
it delete the queue and then recreate as well.

Now imagine this with more than 3k agents...

Change-Id: I762bf2839482ee06c5b2fc9fa50f38f5542cbe99
Closes-bug: #2133389
Signed-off-by: Arnaud Morin <arnaud.morin@ovhcloud.com>