Deprecate enforce_fips_mode
This option depends on FIPS_mode_set API in openssl, which was removed
in openssl 3.0.0 [1]. The packages in both Ubuntu 24.04 (3.0.13) and
Debian 13 (3.5.13) are newer than 3.0.0 so this interface is no longer
functional.

Since RHEL 9, RHEL requires that FIPS mode should be enabled at host
level, instead of each application level.

[1] https://github.com/openssl/openssl/commit/31b069ecea2c567de22b3874c8e71cc37c921ec9

Change-Id: I059e8861256ea6a41e29bed89d12b6d5b73cf4e2
Signed-off-by: Takashi Kajinami <kajinamit@oss.nttdata.com>